requirements.txt
pandas
pandasai
matplotlib
seaborn
openai
python-dotenv
